Support logseq from master
Logseq hasn't landed in unstable yet.

This commit adds support for fetching it from master. Likely I only need
this for a few days, but the approach is a slightly nicer one anyway for
pulling packages from different flakes.

Note that we also remove the (unused) unison flake at the moment.